+++ packages/runtime/src/model-runner/index.ts
@@ -4,5 +4,8 @@ export * from './inputs'
 export * from './outputs'
 export * from './model-runner'
 
+/** @hidden This is not part of the public API; exposed only for use in testing tools. */
+export * from './model-listing'
+
 /** @hidden This is not part of the public API; exposed only for use in performance testing. */
 export * from './perf'